[Feature] Delete Keyframes / Load Figures
+ new Button (Dummy Image) for Key Deletion
+ new function in JumnpToKeyframe for Key Deletion
+ adjustments to draw railImages new after deleting of 1 key
+ Added code in SceneDataManager for Loading of 3D-Figures. Saved name of FigureElements ist wrong o.O

+++ Assets/Scripts/SceneData.cs
@@ -55,9 +55,9 @@ public class SceneryElement
 public class FigureInstanceElement
 {
     public string name;
-    public float instanceNr;        
+    public int instanceNr;        
     public float moment;
-    public float railStart;         // empty auf der es sein soll
+    public int railStart;         // empty auf der es sein soll
 }
 [System.Serializable]
 public class FigureElement
